Termino de algumas estilizações e refatoramento
