+++ modulo10/bloco24/loteria/src/App.js
@@ -1,37 +1,15 @@
-import axios from "axios"
-import React, { useEffect, useState } from "react"
+import React from "react"
 import Router from "./routes/Router"
-import { BASE_URL } from "./constants/BaseUrl"
 import { BrowserRouter } from "react-router-dom"
-import Main from "./components/Main"
+import GetLotteries  from "./Services/GetLotteries"
 
 
-const App = () => {
+export default function App () {
 
-  const [concursos, setConcursos] = useState([])
-
-  const getConcursos = () => {
-    axios
-      .get(`${BASE_URL}/loterias`)
-
-      .then((res) => {
-        setConcursos(res.data)
-      })
-      .catch((error) => {
-        console.log(error)
-      })
-  }
-
-  useEffect(() => {
-    getConcursos()
-  }, [])
-
-  return (
+    return (
     <BrowserRouter>
-      <Main data={concursos}/>
+      <GetLotteries />
       <Router />
     </BrowserRouter>
   );
 }
-
-export default App
