+++ modulo10/bloco24/loteria/src/Services/GetLotteries.js
@@ -1,0 +1,34 @@
+import axios from "axios"
+import Main from "../components/Main"
+import { BASE_URL } from "../constants/BaseUrl"
+import {
+  useEffect,
+  useState
+} from "react"
+
+export default function GetLotteries () {
+
+  const [concursos, setConcursos] = useState([])
+
+  const getConcursos = () => {
+    axios
+      .get(`${BASE_URL}/loterias`)
+
+      .then((res) => {
+        setConcursos(res.data)
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+  }
+
+  useEffect(() => {
+    getConcursos()
+  }, [])
+
+  return (
+    <>
+      <Main data={concursos}/>
+    </>
+  );
+}
