+++ modulo10/bloco24/loteria/src/components/NumeroSorteio.js
@@ -1,4 +1,4 @@
-import { Lista } from "./NumeroSorteioStyle";
+import { Lista } from "../constants/StyledGlobal";
 
 export default function NumeroSorteio ({ numbers }) {
     return <Lista>{numbers}</Lista>
