--- modulo10/bloco24/loteria/src/components/NumeroSorteioStyle.js
@@ -1,16 +1,0 @@
-import styled from "styled-components";
-
-export const Lista = styled.li`
-  background-color: #fff;
-  border-radius: 50%;
-  padding: 1.5em;
-  margin: 20px;
-  font-size: 1.7rem;
-  font-weight: 600;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  @media screen and (max-width: 900px) {
-    padding: 1em;
-  }
-`
