+++ modulo10/bloco24/loteria/src/pages/LotoManiaPage/LotoMania.js
@@ -39,6 +39,7 @@ export default function Lotomania () {
 
   useEffect(() => {
     getRelacaoConcursos()
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
   useEffect(() => {
