+++ modulo10/bloco24/loteria/src/pages/QuinaPage/Quina.js
@@ -39,6 +39,7 @@ export default function Quina () {
 
   useEffect(() => {
     getRelacaoConcursos()
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
   useEffect(() => {
